Site List layout fixes (#3591)
- Fixed Site List overflow.
- Moved "Enable for PDFs" button to Site List page.
