+++ src/ui/popup/site-list-page/index.tsx
@@ -1,4 +1,5 @@
 import {m} from 'malevic';
+import {isFirefox} from '../../../utils/platform';
 import {ViewProps} from '../types';
 import SiteList from './site-list';
 import CheckButton from '../check-button';
@@ -6,6 +7,9 @@ import CheckButton from '../check-button';
 export default function SiteListPage(props: ViewProps) {
     function onSiteListChange(sites: string[]) {
         props.actions.changeSettings({siteList: sites});
+    }
+    function onInvertPDFChange(checked: boolean) {
+        props.actions.changeSettings({enableForPDF: checked});
     }
     function onEnableForProtectedPages(value: boolean) {
         props.actions.changeSettings({enableForProtectedPages: value});
@@ -22,13 +26,21 @@ export default function SiteListPage(props: ViewProps) {
                 onChange={onSiteListChange}
             />
             <label class="site-list-page__description">Enter website name and press Enter</label>
+            {isFirefox() ? null : <CheckButton
+                checked={props.data.settings.enableForPDF}
+                label="Enable for PDF files"
+                description={props.data.settings.enableForPDF ?
+                    'Enabled for PDF documents' :
+                    'Disabled for PDF documents'}
+                onChange={onInvertPDFChange}
+            />}
             <CheckButton
                 checked={props.data.settings.enableForProtectedPages}
                 onChange={onEnableForProtectedPages}
-                label={'Enable on protected pages'}
+                label={'Enable on restricted pages'}
                 description={props.data.settings.enableForProtectedPages ?
-                    'Will try running on protected pages (browser settings etc)' :
-                    'Will not try running on protected pages (browser settings etc)'}
+                    'You should enable it in browser flags too' :
+                    'Disabled for web store and other pages'}
             />
         </div>
     );
