+++ src/ui/popup/site-list-page/style.less
@@ -14,6 +14,10 @@
         text-align: center;
         white-space: nowrap;
     }
+
+    .check-button {
+        margin-top: @indent-small;
+    }
 }
 
 .site-list {
@@ -23,7 +27,7 @@
     display: flex;
     flex-direction: column;
     justify-content: flex-start;
-    max-height: 20rem;
+    max-height: 13.75rem;
     overflow: hidden;
     width: calc(100% - 2 * @size-border);
 
@@ -64,6 +68,7 @@
             right: 0.125rem;
             top: 0.125rem;
             width: 1.25rem;
+            z-index: 1;
 
             &:hover {
                 background-color: @color-control-active;
